+++ app/src/main/java/com/hedvig/app/feature/adyen/payout/AdyenConnectPayoutActivity.kt
@@ -1,13 +1,9 @@
 package com.hedvig.app.feature.adyen.payout
 
+import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.adyen.checkout.base.model.PaymentMethodsApiResponse
 import com.adyen.checkout.base.model.payments.Amount
 import com.adyen.checkout.core.api.Environment
 import com.adyen.checkout.dropin.DropIn
@@ -16,12 +12,10 @@ import com.hedvig.app.BaseActivity
 import com.hedvig.app.R
 import com.hedvig.app.feature.adyen.AdyenCurrency
 import com.hedvig.app.feature.adyen.AdyenDropInService
-import com.hedvig.app.feature.adyen.payin.AdyenRepository
 import com.hedvig.app.getLocale
 import com.hedvig.app.isDebug
 import com.hedvig.app.util.extensions.makeToast
 import e
-import kotlinx.coroutines.launch
 import org.koin.android.viewmodel.ext.android.viewModel
 
 class AdyenConnectPayoutActivity : BaseActivity(R.layout.fragment_container_activity) {
@@ -72,7 +66,9 @@ class AdyenConnectPayoutActivity : BaseActivity(R.layout.fragment_container_acti
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         super.onActivityResult(requestCode, resultCode, data)
-        finish()
+        if (resultCode == Activity.RESULT_CANCELED) {
+            finish()
+        }
     }
 
     companion object {
@@ -84,21 +80,5 @@ class AdyenConnectPayoutActivity : BaseActivity(R.layout.fragment_container_acti
             Intent(context, AdyenConnectPayoutActivity::class.java).apply {
                 putExtra(CURRENCY, currency)
             }
-    }
-}
-
-abstract class AdyenConnectPayoutViewModel : ViewModel() {
-    protected val _payoutMethods = MutableLiveData<PaymentMethodsApiResponse>()
-    val payoutMethods: LiveData<PaymentMethodsApiResponse> = _payoutMethods
-}
-
-class AdyenConnectPayoutViewModelImpl(
-    private val repository: AdyenRepository
-) : AdyenConnectPayoutViewModel() {
-    init {
-        viewModelScope.launch {
-            val response = runCatching { repository.payoutMethods() }
-            response.getOrNull()?.data?.let { _payoutMethods.postValue(it.availablePayoutMethods.paymentMethodsResponse) }
-        }
     }
 }
